Make memory leak much less likey to occur.
Earlier, when loading the spookfile and a previous instance of spook
existed, the EventHandlers weren't properly cleaned out and might keep
references around indefinitely. Now, when calling stop on spook (which
is done when reloading the Spookfile and creating a new spook instance),
the EventHandlers are cleared as part of the stop methods. Also, the
EventHandlers table isn't updated for any type of event handler until
the handler is actually started.