Have Read just give the callback the fd and remove misguided Stdin.
This ofc puts the responsibility on the user to actually issue a read,
otherwise we'll get events indefinitely.

A side note here is that there was an apparent issue with the Read in that
it seemingly went into some kind of infinite loop where it was readable
continuously (even though it shouldn't be since something just read from it).
The reason for that turned out to be garbage collection. The fd:s we get
from ljsyscall are generally wrapped in a metatype and are protected from
garbage collection that way. If we come up with fd:s ourselves (eg. 0 for
stdin) those fd:s need to be wrapped the same way, otherwise bad things happen(tm).