Improved chdir function.
Chdir has always taken a function in which dir will be changed while
resetting it to what it was before chdir when chdir returns.

Chdir is now improved so that not only success of chdir itself but
also any return values of the given function are returned from chdir.

The given function now also receives the success status of chdir so
that it can take appropriate action depending on the success of the
change dir execution.