Deletions are now valid events in stdin mode.
The main fix here is around editors that do "atomic saves" of
files. On BSD (eg. OS X) this is hard to get right and editors
have different opinions on what constitutes an "atomic save" (in
essence they do similar things but not exactly the same things).

The main "fix" is to just accept delete events as valid events in
stdin mode. With that in place an event is fired for the correct file
but it may be a delete event (as with intellij) even though in the
end the path will be there when intellij moves a tmp file in place.

On BSD it is simply not possible to get as fine grained events on
moves as you can on Linux and on BSD you might miss things in between
as you cannot fully rely on kqueue - you have to list directories and
stat files sometimes - which results in windows of "lost" data.

Since kqueue vnode watching is entirely based on inodes rather than
paths, the watcher has been somewhat restructured to not rely so much
on paths.

On Linux the event ordering when moving files around has been
corrected. Previously the move events would always go last.