Use nix to generate build buildkite pipelines.
Also add cachix caching as build step. Add a build.nix. We now rely on
a buildkite hook to ensure all nix derivations are present before
running step command. This is so that when the job is parallelized we
still have everything necessary to run any step. Since Nix knows
what's already present this takes no time if nothing needs to be done
really.